A fix for msvc/windows: it does not attempt to long to int64 in so we got unresolved QTest::qCompare<__int64,long>() overload. Now we cast to qulonglong.
svn path=/trunk/KDE/kdepim/kleopatra/; revision=762823

+++ kleopatra/tests/test_verify.cpp
@@ -75,7 +75,7 @@ class VerifyTest : public QObject
 
       GpgME::Signature sig = result.signature( 0 );
       QCOMPARE( sig.summary() & GpgME::Signature::KeyMissing, 0 );
-      QCOMPARE( sig.creationTime(), 1189650248L );
+      QCOMPARE( (qulonglong)sig.creationTime(), (qulonglong)1189650248UL );
       QCOMPARE( sig.validity(), GpgME::Signature::Full );
     }
 };
